refactor: 도메인형 디렉터리 구조로 변경
